requirements.txt
requests==2.31.0
binance-connector==3.1.0
python-dotenv==1.0.0
krakenex==2.1.0